Extract container-related functions to ldp-container.js
